+++ libs/voltage-rtdb/src/traits.rs
@@ -218,14 +218,14 @@ pub trait Rtdb: Send + Sync + 'static {
     }
 
     /// Enqueue control command to per-channel TODO queue: comsrv:{channel}:C:TODO
-    async fn enqueue_control(&self, channel_id: u16, payload_json: &str) -> Result<()> {
+    async fn enqueue_control(&self, channel_id: u32, payload_json: &str) -> Result<()> {
         let key = ChannelRedisKeys::control_todo(channel_id);
         self.list_rpush(&key, Bytes::from(payload_json.to_string()))
             .await
     }
 
     /// Enqueue adjustment command to per-channel TODO queue: comsrv:{channel}:A:TODO
-    async fn enqueue_adjustment(&self, channel_id: u16, payload_json: &str) -> Result<()> {
+    async fn enqueue_adjustment(&self, channel_id: u32, payload_json: &str) -> Result<()> {
         let key = ChannelRedisKeys::adjustment_todo(channel_id);
         self.list_rpush(&key, Bytes::from(payload_json.to_string()))
             .await
